+++ src/app.ts
@@ -1,13 +1,13 @@
 import db from "../models/index";
 import express from "express";
-import usersRouter from "./routes/users";
-import animalRouter from "./routes/pets";
-import checkoutRouter from "./routes/checkout";
-import visitor from "./routes/visitor";
-import transactionsRouter from "./routes/transaction";
-import reviewsRouter from "./routes/review";
-import commentRouter from "./routes/comment";
-import adminRouter from "./routes/admin";
+import usersRouter from "./routes/user/user-Routes";
+import animalRouter from "./routes/pet/pet-Routes";
+import checkoutRouter from "./routes/checkout/checkout-Routes";
+import visitor from "./routes/visitor/visitor-Routes";
+import transactionsRouter from "./routes/transaction/transaction-Routes";
+import reviewsRouter from "./routes/review/review-Routes";
+import commentRouter from "./routes/comment/comment-Routes";
+import adminRouter from "./routes/admin/admin-Routes";
 import dotenv from "dotenv";
 import cors from "cors";
 dotenv.config();
